requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
Flask-Limiter==3.8.0
Flask-CORS==4.0.1
Werkzeug==3.0.3
Pillow==10.4.0
PyJWT==2.9.0
python-dotenv==1.0.1
bleach==6.1.0
gunicorn==23.0.0
